Extracted functions to module for testability.
They could have stayed in Highlighter.js and been exported, but why
not put them in their own package.
